Reduces webserver and task worker count to 1 by default
